validation ok

+++ shopping_fe/src/components/common/Checkout.js
@@ -1,12 +1,13 @@
 import React, { useState } from 'react';
 import { Container, Stack } from 'react-bootstrap';
 import { Link, Navigate, useNavigate } from 'react-router-dom';
-import { Breadcrumb, Form, Grid, Header, Segment } from 'semantic-ui-react';
+import { Breadcrumb, Form, Grid, Header, Message, Segment } from 'semantic-ui-react';
 import { useAuth } from '../context/AuthContext';
 import { CartItem } from './CartItem';
 import { useShoppingCart } from '../context/ShoppingCartContext';
 import { formatCurrency } from '../utils/formatCurrency';
 import { storeApi } from '../misc/StoreApi';
+import { handleLogError } from '../utils/Helpers';
 
 const Checkout = () => {
     const { getUser } = useAuth()
@@ -27,16 +28,50 @@ const Checkout = () => {
     const [order, setOrder] = useState(initialFormState);
     const navigate = useNavigate();
 
+    const [isError, setIsError] = useState(false);
+    const [errorMessage, setErrorMessage] = useState('');
+    const [nameError, setNameError] = useState('');
+    const [phoneNumError, setPhoneNumError] = useState('');
+    const [addressError, setAddressError] = useState('');
+    const [paymentMethodError, setPaymentMethodError] = useState('');
+
     const handleChange = (event, { name, value }) => {
         setOrder({ ...order, [name]: value })
     }
 
     const handleSubmit = async (event) => {
         event.preventDefault();
+
+        let checkError = false
         await storeApi.addOrder(user, order)
-        removeAllFromCart()
-        await storeApi.deleteListCartItemsByUser(user)
-        navigate('/orders/me');
+        .catch((error) => {
+            checkError = true
+            handleLogError(error);
+            const errorData = error.response;
+            if (errorData.status === 400) {
+                if (errorData.data.fieldErrors) {
+                    errorData.data.fieldErrors.forEach(fieldError => {
+                        if (fieldError.field === 'name') {
+                            setNameError(fieldError.message);
+                        } else if (fieldError.field === 'phoneNum') {
+                            setPhoneNumError(fieldError.message);
+                        } else if (fieldError.field === 'address') {
+                            setAddressError(fieldError.message);
+                        } else if (fieldError.field === 'paymentMethod') {
+                            setPaymentMethodError(fieldError.message);
+                        }
+                    })
+                }
+            } else {
+                setErrorMessage('Lỗi');
+                setIsError(true);
+            }
+        })
+        if (!checkError) {
+            removeAllFromCart()
+            await storeApi.deleteListCartItemsByUser(user)
+            navigate('/orders/me');
+        }
     }
     
     if (!isUser || cartItems === null || cartItems.length < 1) {
@@ -58,8 +93,11 @@ const Checkout = () => {
                                 <Grid.Row>
                                     <Grid.Column width={8}>
                                         <Form.Input fluid className='required' label='Họ tên' name='name' placeholder='Họ tên' value={order.name} onChange={handleChange} />
+                                        { nameError && <span style={{ color: 'red', fontSize: '12px' }}>{nameError}</span> }
                                         <Form.Input fluid className='required' label='Điện thoại' name='phoneNum' placeholder='Điện thoại' value={order.phoneNum} onChange={handleChange} />
+                                        { phoneNumError && <span style={{ color: 'red', fontSize: '12px' }}>{phoneNumError}</span> }
                                         <Form.Input fluid className='required' label='Địa chỉ' name='address' placeholder='Địa chỉ' value={order.address} onChange={handleChange} />
+                                        { addressError && <span style={{ color: 'red', fontSize: '12px' }}>{addressError}</span> }
                                         <Form.Select
                                             fluid
                                             name='paymentMethod'
@@ -70,6 +108,7 @@ const Checkout = () => {
                                             onChange={handleChange}
                                             required
                                         />
+                                        { paymentMethodError && <span style={{ color: 'red', fontSize: '12px' }}>{paymentMethodError}</span> }
                                     </Grid.Column>
                                     <Grid.Column width={8} style={{ maxHeight: '300px', overflowY: 'auto' }} >
                                         <Stack gap={3}>
@@ -82,6 +121,7 @@ const Checkout = () => {
                                 <Grid.Row>
                                     <Grid.Column width={8}>
                                         <Form.Button primary type='submit' floated='right' onClick={() => { }}>Xác nhận</Form.Button>
+                                        {isError && <Message negative>{errorMessage}</Message>}
                                     </Grid.Column>
                                     <Grid.Column width={6}>
                                         <Header as='h4'>
